comic-genai/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
python-dotenv==1.0.1
openai==1.30.4
jinja2==3.1.4
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
